requirements.txt
openai==0.27.10
aiogram==2.25.1
python-dotenv==1.0.0
APScheduler==3.10.4
vosk==0.3.45
googletrans==3.1.0a0